refactor: Separate unit tests
